[c10d] Reduce test time by reusing ProcessGroup
[ghstack-poisoned]
